[minerva] add card button
